Re-factored the code to make it object oriented. Fixed infinite chrome driver spawn.
